exb_transport.f90:   1,nl --> 2,nlm1 flux update
 misc_mod.f90: blon0pt*8 --> blon0pt

 nrlmsise00.f: svv(1) --> svv(25)

 photo_production_mod.f90: 7 --> nneut

+++ misc_mod.f90
@@ -8,7 +8,7 @@ module misc_mod
             u4(nz,nf,nl),u5(nz,nf,nl)
 
     real :: ppt(nzp1,nfp1,nlt)
-    real*8 :: blonp0t(nlt+3)
+    real :: blonp0t(nlt+3)
 
     real :: deni_mnp(nz,nion),ti_mnp(nz,nion),te_mnp(nz)
